src/requirements.txt
Django
Pillow
pyaxmlparser
biplist
markdown
names
psycopg
python-dotenv
